Pin versions of all Python dependencies
I am completely sick of people coming to me and saying "my build used to
work, but now it fails, and all I did was rebuild" and having to track
this down to major changes in un-pinned dependencies.

So, since I don't want to deal with this anymore, I'll just pin all of
these dependencies at the version they would get if the container were
to be built at the present moment.

The author now can update as necessary, without any risk of getting
unintended version upgrades.